third_party/blink/web_tests/http/tests/security/mixedContent/insecure-image-in-main-frame-blocked.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.setPopupBlockingEnabled(false);
    testRunner.overridePreference("WebKitStrictMixedContentChecking", true);
}

window.addEventListener("message", function (e) {
  if (window.testRunner)
      testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window that loads an insecure image.  We should not
trigger a mixed content callback because we've set the preference to block all
mixed content, including passive content.</p>
<script>
window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image.html");
</script>
</body>
</html>
